Fix dumpy query

+++ pkg/api/messagescan.go
@@ -13,7 +13,7 @@ var messageScanMap = map[string]func(incomingChannel string, user string, permis
 		if strings.Contains(strings.ToLower(strings.Join(brokenMessage, " ")), "dumpy") {
 			var dumpyCount int64 = 0
 			// channel_id 2 is #sangnope
-			err := apidb.DB.QueryRow(`UPDATE channel_data SET data = (data::text::integer + 1)::text::json WHERE channel_id = 2 and id = '!dumpy' RETURNING data`).Scan(&dumpyCount)
+			err := apidb.DB.QueryRow(`UPDATE channel_data SET data = (data::text::bigint + 1)::text::json WHERE channel_id = 2 and id = '!dumpy' RETURNING data`).Scan(&dumpyCount)
 			if err != nil {
 				*msgChan <- chat("NO DUMPY!?!?!?! ericareiShock2", incomingChannel)
 				log.Println(err)
